shardtree: Move error types into module
