+++ shardtree/src/prunable.rs
@@ -1,6 +1,4 @@
 use std::collections::{BTreeMap, BTreeSet};
-use std::fmt;
-use std::ops::Range;
 use std::sync::Arc;
 
 use bitflags::bitflags;
@@ -9,6 +7,7 @@ use incrementalmerkletree::{
 };
 use tracing::trace;
 
+use crate::error::{InsertionError, QueryError};
 use crate::{LocatedTree, Node, Tree};
 
 bitflags! {
@@ -324,98 +323,6 @@ pub struct IncompleteAt {
     /// [`MARKED`]: RetentionFlags::MARKED
     pub required_for_witness: bool,
 }
-
-/// An error prevented the insertion of values into the subtree.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum InsertionError {
-    /// The caller attempted to insert a subtree into a tree that does not contain
-    /// the subtree's root address.
-    NotContained(Address),
-    /// The start of the range of positions provided for insertion is not included
-    /// in the range of positions within this subtree.
-    OutOfRange(Position, Range<Position>),
-    /// An existing root hash conflicts with the root hash of a node being inserted.
-    Conflict(Address),
-    /// An out-of-order checkpoint was detected
-    ///
-    /// Checkpoint identifiers must be in nondecreasing order relative to tree positions.
-    CheckpointOutOfOrder,
-    /// An append operation has exceeded the capacity of the tree.
-    TreeFull,
-    /// An input data structure had malformed data when attempting to insert a value
-    /// at the given address
-    InputMalformed(Address),
-}
-
-impl fmt::Display for InsertionError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match &self {
-            InsertionError::NotContained(addr) => {
-                write!(f, "Tree does not contain a root at address {:?}", addr)
-            }
-            InsertionError::OutOfRange(p, r) => {
-                write!(
-                    f,
-                    "Attempted insertion point {:?} is not in range {:?}",
-                    p, r
-                )
-            }
-            InsertionError::Conflict(addr) => write!(
-                f,
-                "Inserted root conflicts with existing root at address {:?}",
-                addr
-            ),
-            InsertionError::CheckpointOutOfOrder => {
-                write!(f, "Cannot append out-of-order checkpoint identifier.")
-            }
-            InsertionError::TreeFull => write!(f, "Note commitment tree is full."),
-            InsertionError::InputMalformed(addr) => {
-                write!(f, "Input malformed for insertion at address {:?}", addr)
-            }
-        }
-    }
-}
-
-impl std::error::Error for InsertionError {}
-
-/// Errors that may be returned in the process of querying a [`ShardTree`]
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum QueryError {
-    /// The caller attempted to query the value at an address within a tree that does not contain
-    /// that address.
-    NotContained(Address),
-    /// A leaf required by a given checkpoint has been pruned, or is otherwise not accessible in
-    /// the tree.
-    CheckpointPruned,
-    /// It is not possible to compute a root for one or more subtrees because they contain
-    /// [`Node::Nil`] values at positions that cannot be replaced with default hashes.
-    TreeIncomplete(Vec<Address>),
-}
-
-impl fmt::Display for QueryError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match &self {
-            QueryError::NotContained(addr) => {
-                write!(f, "Tree does not contain a root at address {:?}", addr)
-            }
-            QueryError::CheckpointPruned => {
-                write!(
-                    f,
-                    "The leaf corresponding to the requested checkpoint is not present in the tree."
-                )
-            }
-            QueryError::TreeIncomplete(addrs) => {
-                write!(
-                    f,
-                    "Unable to compute root; missing values for nodes {:?}",
-                    addrs
-                )
-            }
-        }
-    }
-}
-
-impl std::error::Error for QueryError {}
 
 /// Operations on [`LocatedTree`]s that are annotated with Merkle hashes.
 impl<H: Hashable + Clone + PartialEq> LocatedPrunableTree<H> {
@@ -991,10 +898,13 @@ mod tests {
 
     use incrementalmerkletree::{Address, Level, Position};
 
-    use super::{LocatedPrunableTree, PrunableTree, QueryError, RetentionFlags};
-    use crate::tree::{
-        tests::{leaf, nil, parent},
-        LocatedTree,
+    use super::{LocatedPrunableTree, PrunableTree, RetentionFlags};
+    use crate::{
+        error::QueryError,
+        tree::{
+            tests::{leaf, nil, parent},
+            LocatedTree,
+        },
     };
 
     #[test]
